// src/types.ts
/**
 * ```asn
 * LDSSecurityObjectVersion ::= INTEGER  { v0(0), v1(1) }
 * ```
 */
export enum LDSSecurityObjectVersion {
    v0 = 0,
    v1 = 1,
}

/**
 * ```asn
 * DataGroupNumber ::= INTEGER  { v0(0), v1(1), v2(2),
 * v3(3), v4(4), v5(5), v6(6), v7(7), v8(8), v9(9), v10(10),
 * v11(11), v12(12), v13(13), v14(14), v15(15), v16(16)}
 * ```
 */
export enum DataGroupNumber {
    dataGroup1 = 1,
    dataGroup2 = 2,
    dataGroup3 = 3,
    dataGroup4 = 4,
    dataGroup5 = 5,
    dataGroup6 = 6,
    dataGroup7 = 7,
    dataGroup8 = 8,
    dataGroup9 = 9,
    dataGroup10 = 10,    
    dataGroup11 = 11,
    dataGroup12 = 12,
    dataGroup13 = 13,
    dataGroup14 = 14,
    dataGroup15 = 15,
    dataGroup16 = 16,
}